Refactor directory
1) Rearrange the routes into a dedicated directory -> routes
2) Compile all models into models.py
3) Only need to call the __init__ file to launch
