+++ App/__init__.py
@@ -1,26 +1,22 @@
-import os
 from flask import Flask
+from .extensions import db
+from flask_cors import CORS
+from .routes.grant import grant
+from .routes.member import member
+from .routes.household import household
+
 
 def create_app():
-    # create and configure the app
     app = Flask(__name__)
-    app.config.from_mapping(
-        SECRET_KEY='dev',
-        DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
-    )
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:@127.0.0.1:3306/government-grant-disbursement'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-    if test_config is None:
-        # load the instance config, if it exists, when not testing
-        app.config.from_pyfile('config.py', silent=True)
-    else:
-        # load the test config if passed in
-        app.config.from_mapping(test_config)
+    db.init_app(app)
 
-    # ensure the instance folder exists
-    try:
-        os.makedirs(app.instance_path)
-    except OSError:
-        pass
+    app.register_blueprint(member)
+    app.register_blueprint(household)
+    app.register_blueprint(grant)
 
+    CORS(app)
 
     return app
